setup circular link removal and internal ip directions added

+++ setup.sh
@@ -62,9 +62,11 @@ ln -s "$CWD/.tmux.conf" "${HOME}"
 pip install --user git+git://github.com/Lokaltog/powerline 
 POWERLINE_ROOT="$(pip show powerline-status | grep '^Location: ' | sed 's/^Location: //')/powerline"
 powerline_config="$POWERLINE_ROOT/config_files"
-mv $powerline_config $powerline_config.old
-ln -s "$CWD/powerline_config" "$powerline_config"
-ln -s "$POWERLINE_ROOT/bindings/tmux/powerline.conf" "${HOME}" 
+if [ ! -L $powerline_config ]; then
+    mv $powerline_config $powerline_config.old
+    ln -s "$CWD/powerline_config" "$powerline_config"
+    ln -s "$POWERLINE_ROOT/bindings/tmux/powerline.conf" "${HOME}" 
+fi
 
 # install powerline fonts
 if [[ ! -d ${HOME}/.powerline_fonts ]]; then
